+++ Dalamud.Boot/Dalamud.Boot.vcxproj
@@ -93,7 +93,9 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">NotUsing</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">NotUsing</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="budget_hooks.cpp" />
+    <ClCompile Include="unicode.cpp" />
+    <ClCompile Include="xivfixes.cpp" />
+    <ClCompile Include="utils.cpp" />
     <ClCompile Include="pch_nmd_assembly_impl.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">NotUsing</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">NotUsing</PrecompiledHeader>
@@ -115,7 +117,10 @@
     <ClInclude Include="..\lib\CoreCLR\core\coreclr_delegates.h" />
     <ClInclude Include="..\lib\CoreCLR\core\hostfxr.h" />
     <ClInclude Include="..\lib\CoreCLR\nethost\nethost.h" />
-    <ClInclude Include="budget_hooks.h" />
+    <ClInclude Include="bootconfig.h" />
+    <ClInclude Include="hooks.h" />
+    <ClInclude Include="unicode.h" />
+    <ClInclude Include="utils.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="veh.h" />
   </ItemGroup>
