+++ Dalamud.Boot/Dalamud.Boot.vcxproj.filters
@@ -28,14 +28,20 @@
     <ClCompile Include="..\lib\CoreCLR\boot.cpp">
       <Filter>CoreCLR</Filter>
     </ClCompile>
-    <ClCompile Include="budget_hooks.cpp">
-      <Filter>Dalamud.Boot DLL</Filter>
-    </ClCompile>
     <ClCompile Include="pch_nmd_assembly_impl.cpp">
       <Filter>Project Files</Filter>
     </ClCompile>
     <ClCompile Include="pch.cpp">
       <Filter>Project Files</Filter>
+    </ClCompile>
+    <ClCompile Include="utils.cpp">
+      <Filter>Dalamud.Boot DLL</Filter>
+    </ClCompile>
+    <ClCompile Include="xivfixes.cpp">
+      <Filter>Dalamud.Boot DLL</Filter>
+    </ClCompile>
+    <ClCompile Include="unicode.cpp">
+      <Filter>Dalamud.Boot DLL</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -57,11 +63,20 @@
     <ClInclude Include="veh.h">
       <Filter>Dalamud.Boot DLL</Filter>
     </ClInclude>
-    <ClInclude Include="budget_hooks.h">
-      <Filter>Dalamud.Boot DLL</Filter>
-    </ClInclude>
     <ClInclude Include="pch.h">
       <Filter>Project Files</Filter>
+    </ClInclude>
+    <ClInclude Include="utils.h">
+      <Filter>Dalamud.Boot DLL</Filter>
+    </ClInclude>
+    <ClInclude Include="hooks.h">
+      <Filter>Dalamud.Boot DLL</Filter>
+    </ClInclude>
+    <ClInclude Include="unicode.h">
+      <Filter>Dalamud.Boot DLL</Filter>
+    </ClInclude>
+    <ClInclude Include="bootconfig.h">
+      <Filter>Dalamud.Boot DLL</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
