+++ Dalamud.Boot/dllmain.cpp
@@ -1,83 +1,34 @@
 #include "pch.h"
 
-#include "budget_hooks.h"
+#include "bootconfig.h"
+#include "hooks.h"
 #include "veh.h"
 
 HMODULE g_hModule;
 HINSTANCE g_hGameInstance = GetModuleHandleW(nullptr);
-
-bool check_env_var(std::string name) {
-    size_t required_size;
-    getenv_s(&required_size, nullptr, 0, name.c_str());
-    if (required_size > 0) {
-        if (char* is_no_veh = static_cast<char*>(malloc(required_size * sizeof(char)))) {
-            getenv_s(&required_size, is_no_veh, required_size, name.c_str());
-            auto result = _stricmp(is_no_veh, "true");
-            free(is_no_veh);
-            if (result == 0)
-                return true;
-        }
-    }
-
-    return false;
-}
-
-bool is_running_on_linux() {
-    size_t required_size;
-    getenv_s(&required_size, nullptr, 0, "XL_WINEONLINUX");
-    if (required_size > 0) {
-        if (char* is_wine_on_linux = static_cast<char*>(malloc(required_size * sizeof(char)))) {
-            getenv_s(&required_size, is_wine_on_linux, required_size, "XL_WINEONLINUX");
-            auto result = _stricmp(is_wine_on_linux, "true");
-            free(is_wine_on_linux);
-            if (result == 0)
-                return true;
-        }
-    }
-
-    HMODULE hntdll = GetModuleHandleW(L"ntdll.dll");
-    if (!hntdll) // not running on NT
-        return true;
-
-    FARPROC pwine_get_version = GetProcAddress(hntdll, "wine_get_version");
-    FARPROC pwine_get_host_version = GetProcAddress(hntdll, "wine_get_host_version");
-
-    return pwine_get_version != nullptr || pwine_get_host_version != nullptr;
-}
-
-bool is_veh_enabled() {
-    return check_env_var("DALAMUD_IS_VEH");
-}
-
-bool is_full_dumps() {
-    return check_env_var("DALAMUD_IS_VEH_FULL");
-}
 
 DllExport DWORD WINAPI Initialize(LPVOID lpParam, HANDLE hMainThreadContinue) {
 #ifndef NDEBUG
     ConsoleSetup(L"Dalamud Boot");
 #endif
     try {
-        budget_hooks::fixes::apply_all(true);
+        xivfixes::apply_all(true);
     } catch (const std::exception& e) {
-        MessageBoxA(nullptr, std::format("Failed to do general fixups. Some things might not work.\n\nDescription: {}", e.what()).c_str(), "Dalamud Boot", MB_OK);
+        std::cerr << std::format("Failed to do general fixups. Some things might not work.\nError: {}\n", e.what());
     }
 
-    printf("Dalamud.Boot Injectable, (c) 2021 XIVLauncher Contributors\nBuilt at: %s@%s\n\n", __DATE__, __TIME__);
+    std::cerr << "Dalamud.Boot Injectable, (c) 2021 XIVLauncher Contributors\nBuilt at: " __DATE__ "@" __TIME__ "\n\n";
 
-    if (check_env_var("DALAMUD_WAIT_DEBUGGER")) {
-        printf("Waiting for debugger to attach...\n");
+    if (bootconfig::is_wait_debugger()) {
+        std::cerr << "Waiting for debugger to attach...\n";
         while (!IsDebuggerPresent())
             Sleep(100);
-        printf("Debugger attached.\n");
+        std::cerr << "Debugger attached.\n";
     }
 
-    wchar_t _module_path[MAX_PATH];
-    GetModuleFileNameW(g_hModule, _module_path, sizeof _module_path / 2);
-    std::filesystem::path fs_module_path(_module_path);
-
-    std::wstring runtimeconfig_path = _wcsdup(fs_module_path.replace_filename(L"Dalamud.runtimeconfig.json").c_str());
-    std::wstring module_path = _wcsdup(fs_module_path.replace_filename(L"Dalamud.dll").c_str());
+    const auto fs_module_path = utils::get_module_path(g_hModule);
+    const auto runtimeconfig_path = std::filesystem::path(fs_module_path).replace_filename(L"Dalamud.runtimeconfig.json").wstring();
+    const auto module_path = std::filesystem::path(fs_module_path).replace_filename(L"Dalamud.dll").wstring();
 
     // ============================== CLR ========================================= //
 
@@ -94,27 +45,28 @@ DllExport DWORD WINAPI Initialize(LPVOID lpParam, HANDLE hMainThreadContinue) {
     if (result != 0)
         return result;
 
-    typedef void (CORECLR_DELEGATE_CALLTYPE* custom_component_entry_point_fn)(LPVOID, HANDLE);
-    custom_component_entry_point_fn entrypoint_fn = reinterpret_cast<custom_component_entry_point_fn>(entrypoint_vfn);
+    using custom_component_entry_point_fn = void (CORECLR_DELEGATE_CALLTYPE*)(LPVOID, HANDLE);
+    const auto entrypoint_fn = reinterpret_cast<custom_component_entry_point_fn>(entrypoint_vfn);
 
     // ============================== VEH ======================================== //
 
-    printf("Initializing VEH... ");
-    if (is_running_on_linux()) {
-        printf("VEH was disabled, running on linux\n");
-    } else if (is_veh_enabled()) {
-        if (veh::add_handler(is_full_dumps()))
-            printf("Done!\n");
-        else printf("Failed!\n");
+    std::cerr << "Initializing VEH...\n";
+    if (utils::is_running_on_linux()) {
+        std::cerr << "=> VEH was disabled, running on linux\n";
+    } else if (bootconfig::is_veh_enabled()) {
+        if (veh::add_handler(bootconfig::is_veh_full()))
+            std::cerr << "=> Done!\n";
+        else
+            std::cerr << "=> Failed!\n";
     } else {
-        printf("VEH was disabled manually\n");
+        std::cerr << "VEH was disabled manually\n";
     }
 
     // ============================== Dalamud ==================================== //
 
-    printf("Initializing Dalamud... ");
+    std::cerr << "Initializing Dalamud... ";
     entrypoint_fn(lpParam, hMainThreadContinue);
-    printf("Done!\n");
+    std::cerr << "Done!\n";
 
 #ifndef NDEBUG
     fclose(stdin);
@@ -134,7 +86,7 @@ BOOL APIENTRY DllMain(const HMODULE hModule, const DWORD dwReason, LPVOID lpRese
             g_hModule = hModule;
             break;
         case DLL_PROCESS_DETACH:
-            budget_hooks::fixes::apply_all(false);
+            xivfixes::apply_all(false);
             veh::remove_handler();
             break;
     }
